Add file move and rename prediction, other fixes
If a txt file is renamed or moved to another folder outside of the app (not both), the app will detect it and prompt the user to verify changes. Tags will be migrated from the old Textfile instance to the new one if the user indicates that the "new" file is simply the old one moved/renamed. This is done whenever the home page is requested and changes to the files are found.
There will be false positives (e.g. one file deleted and one file created in the same directory): that's why changes are to be verified.
Verification is done with a popup. Popup has one page per possible tag transfer. User can choose to transfer tags or ignore; either way, the old Textfile instance will be deleted.
The purpose of this feature is to make it so that tags are not always lost if files renamed/moved outside of the app (e.g. in the OS filesystem); it's a safeguard measure.

Also fixed a div nesting problem in User's new view, and renamed some variables.