Add Textfile archiving, temp fix for filepath case-insensitivity issue, bug/minor fixes
Previously, if the home directory was changed, all Textfiles not in the new home were lost. Now, they will only be hidden from the home page (i.e. archived). They will still be accessible from tag pages, or by manually going to url. To unarchive, change the home directory in Preferences.

Removed StringConstructor#directory_valid? because it's identical to Dir#exists?, so is not needed.
Generalized the User model's pre-save method that removed redundant slashes in directory filepath.

"Fixed" a bug where setting a home directory was case sensitive. Filepaths in Windows are case insensitive, so C:\Program Files should be the same as C:\program files. To account for this "feature" of Windows, all directory paths are forced to be lowercase. File names are untouched (i.e. the same as whatever they were created as). The location attribute of Textfiles is therefore in the style "c:\users\documents\Collection-of-587-Haikus.txt". The show view has everything lowercase because this style is ugly.

This is a terribly hacky way of avoiding the problem, and the next commit will give it a proper fix.