Add data import/export, major tag- and save-related bug fixes
Data saved in json format; files timestamped. To import data, the home
directory must be the same as the save data's home directory. The newest
data is the one that is imported.

Bug fixes:
- Fixed bug where new tags created in Textfile form not shown in tag
sidebar.
- Fixed bug where new tags created in Create Tag form not shown if added
to existing category.
- Fixed bug where updating Textfile would throw Write To Disk Error
despite no problems with save.
- Fixed bug where Textfile save was missing a check for form data before
setting category for new tags.

This is what I get for committing too early.